owed 8: ordinal reads through a 64-bit temporary
An ordinal width of 8 is admissible (§4.5). The compiled remap read
through uint32_t, so a 64-bit ordinal overflowed the temporary. The
identity clamp on a union tag had the same 32-bit compare; enums
already used uint64_t.

+++ internal/codegen/cpptable/fixedruntime.go
@@ -304,11 +304,11 @@ TABLE_FIXED_INLINE void TableFixedApply( const TableFixedEntry & p, const uint8_
             // enum gained a variant IN THE MIDDLE is remapped here and never
             // reinterpreted. The table is the plan's own, laid down by the
             // compiler above the entries.
-            uint32_t raw = 0;
+            uint64_t raw = 0;
             memcpy( &raw, src + p.src, p.size );
             const uint16_t * table = (const uint16_t *) (const void *) ( base + p.aux );
-            uint32_t v = 0;
-            if ( raw != 0 && raw <= table[0] ) { v = table[raw]; }
+            uint64_t v = 0;
+            if ( raw != 0 && raw <= (uint64_t) table[0] ) { v = table[raw]; }
             memcpy( dst + p.dst, &v, p.dstsize );
             break;
         }
